+++ R/provider-anthropic.R
@@ -293,8 +293,8 @@ method(value_turn, ProviderAnthropic) <- function(
   })
 
   tokens <- value_tokens(provider, result)
-  tokens_log(provider, tokens)
-  assistant_turn(contents, json = result, tokens = unlist(tokens))
+  cost <- get_token_cost(provider, tokens)
+  assistant_turn(contents, json = result, tokens = unlist(tokens), cost = cost)
 }
 
 # ellmer -> Claude --------------------------------------------------------------